+++ src/main/java/com/jcloisterzone/game/phase/ActionPhase.java
@@ -9,19 +9,14 @@ import com.google.common.collect.Iterables;
 import com.jcloisterzone.LittleBuilding;
 import com.jcloisterzone.action.MeepleAction;
 import com.jcloisterzone.action.PlayerAction;
-import com.jcloisterzone.action.TakePrisonerAction;
 import com.jcloisterzone.board.Location;
 import com.jcloisterzone.board.Position;
 import com.jcloisterzone.board.TileTrigger;
 import com.jcloisterzone.board.pointer.FeaturePointer;
-import com.jcloisterzone.board.pointer.MeeplePointer;
 import com.jcloisterzone.event.FlierRollEvent;
 import com.jcloisterzone.event.NeutralFigureMoveEvent;
 import com.jcloisterzone.event.SelectActionEvent;
-import com.jcloisterzone.event.TowerIncreasedEvent;
 import com.jcloisterzone.feature.City;
-import com.jcloisterzone.feature.Tower;
-import com.jcloisterzone.figure.Follower;
 import com.jcloisterzone.figure.Meeple;
 import com.jcloisterzone.figure.SmallFollower;
 import com.jcloisterzone.figure.predicate.MeeplePredicates;
@@ -90,7 +85,7 @@ public class ActionPhase extends Phase {
         }
     }
 
-   
+
 
     @Override
     public void placeTowerPiece(Position p) {
@@ -166,6 +161,7 @@ public class ActionPhase extends Phase {
 
     @WsSubscribe
     public void handleFlierDice(FlierDiceMessage msg) {
+        flierCap.setFlierUsed(true);
         flierCap.setFlierDistance(msg.getMeepleTypeClass(), msg.getDistance());
         game.post(new FlierRollEvent(getActivePlayer(), getTile().getPosition(), msg.getDistance()));
         next(FlierActionPhase.class);